+++ packages/framework/src/Foundation/FileCollection.php
@@ -7,13 +7,13 @@ namespace Hyde\Foundation;
 use Hyde\Foundation\Concerns\BaseFoundationCollection;
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Helpers\Features;
-use Hyde\Framework\Models\Support\File;
 use Hyde\Framework\Services\DiscoveryService;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\HtmlPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\File;
 
 /**
  * @see \Hyde\Framework\Foundation\FileCollection
