+++ packages/framework/src/Framework/Services/DiscoveryService.php
@@ -6,12 +6,12 @@ namespace Hyde\Framework\Services;
 
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Exceptions\UnsupportedPageTypeException;
-use Hyde\Framework\Models\Support\File;
 use Hyde\Hyde;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\File;
 
 /**
  * The core service that powers all HydePHP file auto-discovery.
