+++ packages/framework/src/Framework/Services/SitemapService.php
@@ -7,12 +7,12 @@ declare(strict_types=1);
 namespace Hyde\Framework\Services;
 
 use Exception;
-use Hyde\Framework\Models\Support\Route;
 use Hyde\Hyde;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\Route;
 use SimpleXMLElement;
 
 /**
