+++ packages/framework/src/Hyde.php
@@ -10,7 +10,7 @@ use Hyde\Foundation\PageCollection;
 use Hyde\Foundation\RouteCollection;
 use Hyde\Framework\Concerns\HydePage;
 use Hyde\Framework\Helpers\Features;
-use Hyde\Framework\Models\Support\Route;
+use Hyde\Support\Route;
 use Illuminate\Support\Facades\Facade;
 
 /**
