+++ packages/framework/src/Navigation/NavItem.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Hyde\Navigation;
 
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Framework\Models\Support\Route;
 use Hyde\Hyde;
+use Hyde\Support\Route;
 use Illuminate\Support\Str;
 use Stringable;
 
@@ -30,7 +30,7 @@ class NavItem implements Stringable
     /**
      * Create a new navigation menu item.
      *
-     * @param  \Hyde\Framework\Models\Support\Route|null  $route
+     * @param  \Hyde\Support\Route|null  $route
      * @param  string  $label
      * @param  int  $priority
      * @param  bool  $hidden
