+++ packages/framework/src/Pages/DocumentationPage.php
@@ -7,7 +7,7 @@ namespace Hyde\Pages;
 use Hyde\Framework\Actions\GeneratesSidebarTableOfContents;
 use Hyde\Framework\Concerns\BaseMarkdownPage;
 use Hyde\Framework\Contracts\FrontMatter\DocumentationPageSchema;
-use Hyde\Framework\Models\Support\Route;
+use Hyde\Support\Route;
 
 /**
  * Page class for documentation pages.
