+++ packages/framework/src/Support/DateString.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Models\Support;
+namespace Hyde\Support;
 
 use DateTime;
 use Stringable;
