+++ packages/framework/src/Support/Site.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Models\Support;
+namespace Hyde\Support;
 
 use Hyde\Metadata\GlobalMetadataBag;
