+++ packages/framework/tests/Feature/BuildOutputDirectoryCanBeChangedTest.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Framework\HydeServiceProvider;
-use Hyde\Framework\Models\Support\Site;
 use Hyde\Framework\Services\RebuildService;
 use Hyde\Hyde;
+use Hyde\Support\Site;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\File;
