+++ packages/framework/tests/Feature/FileCollectionTest.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Foundation\FileCollection;
-use Hyde\Framework\Models\Support\File;
 use Hyde\Hyde;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\File;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
