+++ packages/framework/tests/Feature/FileTest.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use Hyde\Framework\Models\Support\File;
 use Hyde\Hyde;
 use Hyde\Pages\MarkdownPage;
+use Hyde\Support\File;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Framework\Models\Support\File
+ * @covers \Hyde\Support\File
  */
 class FileTest extends TestCase
 {
