+++ packages/framework/tests/Feature/HydeKernelTest.php
@@ -7,12 +7,12 @@ namespace Hyde\Framework\Testing\Feature;
 use Composer\InstalledVersions;
 use Hyde\Foundation\HydeKernel;
 use Hyde\Framework\Helpers\Features;
-use Hyde\Framework\Models\Support\Route;
 use Hyde\Hyde;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\Route;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\View;
