+++ packages/framework/tests/Feature/NavigationMenuTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use Hyde\Framework\Models\Support\Route;
 use Hyde\Hyde;
 use Hyde\Navigation\NavigationMenu;
 use Hyde\Navigation\NavItem;
+use Hyde\Support\Route;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
