+++ packages/framework/tests/Feature/RouteCollectionTest.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Foundation\RouteCollection;
-use Hyde\Framework\Models\Support\Route;
 use Hyde\Hyde;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\Route;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
