+++ packages/framework/tests/Feature/StaticSiteServiceTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use Hyde\Framework\Models\Support\Site;
 use Hyde\Hyde;
+use Hyde\Support\Site;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\File;
