+++ packages/framework/tests/Unit/DateStringTest.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Unit;
 
 use DateTime;
-use Hyde\Framework\Models\Support\DateString;
+use Hyde\Support\DateString;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Hyde\Framework\Models\Support\DateString
+ * @covers \Hyde\Support\DateString
  */
 class DateStringTest extends TestCase
 {
