+++ packages/framework/tests/Unit/HydeServiceProviderTest.php
@@ -6,13 +6,13 @@ namespace Hyde\Framework\Testing\Unit;
 
 use Hyde\DataCollections\DataCollectionServiceProvider;
 use Hyde\Framework\HydeServiceProvider;
-use Hyde\Framework\Models\Support\Site;
 use Hyde\Framework\Services\AssetService;
 use Hyde\Hyde;
 use Hyde\Pages\BladePage;
 use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
+use Hyde\Support\Site;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\Artisan;
