+++ packages/framework/tests/Unit/NavItemTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Unit;
 
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Framework\Models\Support\Route;
 use Hyde\Navigation\NavItem;
+use Hyde\Support\Route;
 use Hyde\Testing\TestCase;
 
 /**
