+++ packages/framework/tests/Unit/Views/Components/LinkComponentTest.php
@@ -20,7 +20,7 @@ class LinkComponentTest extends TestCase
 
     public function test_link_component_can_be_rendered_with_route()
     {
-        $route = \Hyde\Framework\Models\Support\Route::get('index');
+        $route = \Hyde\Support\Route::get('index');
         $this->assertEquals('<a href="index.html">bar</a>', rtrim(
             Blade::render('<x-link href="'.$route.'">bar</x-link>')));
     }
@@ -28,7 +28,7 @@ class LinkComponentTest extends TestCase
     public function test_link_component_can_be_rendered_with_route_for_nested_pages()
     {
         View::share('currentPage', 'foo/bar');
-        $route = \Hyde\Framework\Models\Support\Route::get('index');
+        $route = \Hyde\Support\Route::get('index');
         $this->assertEquals('<a href="../index.html">bar</a>', rtrim(
             Blade::render('<x-link href="'.$route.'">bar</x-link>')));
     }
